requirements.txt
# Neuro-Mamba Shielded Control Agent — Pinned Dependencies
# Python 3.10 required for mamba-ssm compatibility

# Core Deep Learning
torch==2.1.0
torchvision
torchaudio

# Mamba SSM (requires CUDA 11.8+ and gcc>=9)
mamba-ssm==1.2.0
causal-conv1d>=1.1.0

# Safety Environments
safety-gymnasium==0.4.0
gymnasium==0.29.0

# Optimization Solvers (CBF-QP Shield)
cvxpy==1.4.2
osqp==0.6.5
qpth                       # differentiable QP layer

# Baselines (PPO-MLP, PPO-LSTM)
stable-baselines3==2.2.1
sb3-contrib                 # RecurrentPPO for LSTM baseline

# Visualization & Logging
wandb
matplotlib
imageio
imageio-ffmpeg              # for MP4 encoding
pygame
opencv-python

# Config
pyyaml
